fix: switch to remote couch api
