--- tests/src/browser/pouch-adapter-test.js
@@ -1,6 +1,0 @@
-const test = require('../../smalltest')
-let api
-
-test('api constructor test setup', async t => {
-  t.ok(true, 'tests work')
-})
